feat(refactor): Move AI prompts into new file (#181)
* separate ai prompts into different file

* move ai dirs
