+++ server/ai/types.ts
